# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API - GES-CARS-2026",
        "description": "API RESTful pour la gestion de location de voitures",
        "contact": {
            "email": "admin@ges-cars.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1"
        }
    ],
    "paths": {
...
        "/insurances": {
            "get": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Liste des assurances",
                "operationId": "f114a2c9f1197bf89f1ac1c463b1aa58",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "third_party",
                                "comprehensive",
                                "all_risk"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Créer une assurance",
                "operationId": "dd244bd9542a307c224dcb8c58ea2879",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "insurance_company",
                                    "policy_number",
                                    "type",
                                    "start_date",
                                    "end_date",
                                    "premium_amount"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "insurance_company": {
                                        "type": "string"
                                    },
                                    "policy_number": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "third_party",
                                            "comprehensive",
                                            "all_risk"
                                        ]
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "premium_amount": {
                                        "type": "number"
                                    },
                                    "deductible_amount": {
                                        "type": "number"
                                    },
                                    "agent_name": {
                                        "type": "string"
                                    },
                                    "agent_phone": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Insurance created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/{id}": {
            "get": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Détails d'une assurance",
                "operationId": "ad29355420cd891aa5c8c5a8f3e0ee8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Modifier une assurance",
                "operationId": "3c008ec39c51fc5d509ee5593ad5fc87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Supprimer une assurance",
                "operationId": "e2e69136f646ca01c1f6b01bf84b2bd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/expired": {
            "get": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Assurances expirées",
                "operationId": "50c330aa338883a70ef5b3301379d522",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/expiring-soon": {
            "get": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Assurances expirant bientôt",
                "operationId": "e9d492db4df072cbb89c14291262fb60",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/{id}/policy-document": {
            "post": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Upload police d'assurance",
                "operationId": "e25dfe5dc6e2ca013f8824c9c78e49a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "policy_document": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/{id}/green-card": {
            "post": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Upload carte verte",
                "operationId": "263050f7e1501aab2d22070196d7a04e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "green_card": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/{id}/attachments": {
            "post": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Upload pièces jointes",
                "operationId": "7db00189612f0a14e074b4c0169fe527",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "attachments[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/{id}/media/{mediaId}": {
            "delete": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Supprimer un média",
                "operationId": "82a6773ca7c6ebb0bb4410f025392d29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mediaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/insurances/{id}/media": {
            "get": {
                "tags": [
                    "Insurances"
                ],
                "summary": "Lister les médias d'une assurance",
                "operationId": "75ed32a28dd4683164ba558d96141a6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrer le token JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}